Simple API Integration

+++ src/MyBudget.Api/Services/IBudgetService.cs
@@ -4,7 +4,7 @@ namespace MyBudget.Api.Services
 {
 	public interface IBudgetService
 	{
-		Task<bool> Add(int value, string desc);
-		Task<bool> Remove(int value, string desc);
+		Task<bool> Add(double value, string desc);
+		Task<bool> Remove(double value, string desc);
 	}
 }
